requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==5.0.2
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements.in
jinja2==3.1.3
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
pillow==10.2.0
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pytz==2024.1
    # via djangorestframework
sqlparse==0.4.4
    # via django
tzdata==2023.4
    # via django
